# backendPI4/api/openapi.yaml
openapi: 3.0.0
info:
  title: API para a aplicação ViriatosScouting
  description:
    "Este documento apresenta toda a documentação de uma API que permite\
    \ gerir um sistema que faz gestão de uma equipa de scouting. A API inclui funcionalidades\
    \ que permitem o utilizador gerir jogadores, clubes, escalões, equipas e utilizadores.\
    \ A mesma tambem apresenta um local para criar tarefas e relatórios."
  version: "1.0"
servers:
  - url: https://localhost:8080
    description: Servidor local utilizado para testar a API
  - url: https://backendpi4-bbzf.onrender.com
    description: Servidor online utilizado para testar a API
...
paths:
...
  /uploads/{filename}:
    get:
      tags:
        - Uploads
      summary: Buscar imagem carregada
      operationId: buscarImagem
      description: Retorna uma imagem guardada no servidor.
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
          description: Nome do ficheiro a ser buscado.
      responses:
        "200":
          description: Imagem encontrada com sucesso
          content:
            image/png:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
        "404":
          description: Ficheiro não encontrado
      security:
        - bearerAuth: []
      x-swagger-router-controller: Upload
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT